Reorganizar definiciones de CountryStructure Fixes #15
