+++ src/CPIDataBase/src/utils/varinteran.jl
